feat: renamed pages to features
